# shared/openapi/openapi.yaml
# The contract. This file is the source of truth for apps/backend's REST
# API - authored and versioned by hand, not generated from the FastAPI
# route code. It lives in shared/ and is mounted read-only by every module
# that needs it (apps/backend, apps/frontend's /api-specs, Specmatic,
# agentgateway) - none of them has to go through the backend to read it.
# apps/backend loads it at startup and serves it verbatim at GET
# /openapi.json (see app/main.py) instead of letting FastAPI derive a
# schema from its own routes. Its examples for Specmatic (the ones that
# need a bearer token or an error response) live in examples/ next to it.
#
# This is a deliberate reversal of this repo's earlier "no checked-in
# schema, always live-generated" stance: a schema generated FROM the
# implementation can never structurally disagree with it, so a provider
# verification test (Specmatic) run against it can only ever catch
# behavioral bugs, not real contract drift. With a physically separate,
# hand-maintained contract, "does the implementation still honor this
# file" becomes a real, failable question - which is the whole point of
# Contract-Driven Development. Consumers (apps/frontend, kafka-bridge,
# Specmatic's mock for consumer-side testing) and the Provider
# (apps/backend) both build against this one file independently.
#
# GraphQL (/graphql) is deliberately not described here - it's a separate
# protocol with its own schema (introspectable at /graphql itself), and
# OpenAPI can't describe its request shape ({query, variables}) in any
# meaningful way. See AGENTS.md's Specmatic section for the full reasoning.
#
# Keep this in sync by hand whenever app/routers/*.py or app/schemas.py
# change - that's the tradeoff for the contract actually meaning something.
openapi: 3.0.3
info:
  title: NASEBANAL Quickstarts API
  # English, not localized: OpenAPI has no per-field i18n mechanism (every
  # description/summary is a single string), and Scalar's own UI chrome is
  # deliberately pinned to English too (see ScalarDoc.tsx) - keeping this
  # spec's own prose in the same language avoids a spec that's part
  # English (schema titles, operation summaries, tags), part Japanese (only
  # this field). Maintaining a second, locale-specific copy of this file
  # was considered and rejected: this is the hand-maintained, single
  # source-of-truth contract (see the header comment above) - a second
  # copy would only ever differ in this one field, but the two would still
  # need to be kept in lockstep for every real schema/path change,
  # doubling the maintenance cost of the one property (no drift) this
  # file's whole design exists for.
  description: >-
    The test-target app's backend API (a minimal event-sourced
    implementation). /transactions is the append-only event log itself;
    /accounts is the balance view derived from it. This contract
    (openapi.yaml) is the source of truth, and the implementation follows it.
  version: 0.1.0
# Every operation needs a bearer token (an RS256 JWT, see /auth/login) unless it says `security: []`
# itself: the login, the health check and the JWKS that verifies the tokens.
security:
  - HTTPBearer: []
paths:
  /health:
    get:
      tags: [health]
      summary: Health
      operationId: health_health_get
      security: []
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              examples:
                healthy:
                  value:
                    status: ok
  /auth/login:
    post:
      tags: [auth]
      summary: Login
      operationId: login_auth_login_post
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              # Named to match the response example below - request and
              # response examples are paired by matching key, not by content.
              demo_login:
                summary: Log in as the seeded demo user
                value:
                  username: demo
                  password: demo
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              examples:
                demo_login:
                  value:
                    # Matches the demo token in authenticated external examples.
                    token: eyJhbGciOiJSUzI1NiIsImtpZCI6Im1YZmJOUEFmS0M0TUN1ZDAiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwODAiLCJzdWIiOiJkZW1vIiwicHJlZmVycmVkX3VzZXJuYW1lIjoiZGVtbyIsImF1ZCI6Im5iLXF1aWNrc3RhcnRzLWFwaSIsImlhdCI6MTc2NzIyNTYwMCwiZXhwIjo0MTAyNDQ0ODAwfQ.egX-L8FA_KK2GhL7Gsg2mYtsWDhmxDxCz6P1h7jUMSAu5xU2oTb5fbib4vBVpay-ExfdBbOvyJ2qDx3M2t2XdrqsdkyWo-GVZFgXCBLVjPY_8s8-CGaVSnhOJMJ2bZOsGHnMTd9ay5LrKw5hHTVAFK1V0Gf9Sk3eqQNIkk987IhEi6b7g2pWE8AS9BrVt9RbRSANo8I_iWVPHpQCF-KsTizgHciTuHvDS0viGa-w_4N8o3nXSMIeTxhuzVgOAJKfccDYhMpUZXc6m0Sco0sxNkGY-J-5p1FEAmEaZ5r_mqWvPKm3xXJ1cr1L5WiJrUY9GyFs2uN0P6T-Jn4ynBpA5Q
                    username: demo
        "401":
          description: Unknown username or wrong password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              examples:
                bad_credentials:
                  value:
                    detail: invalid username or password
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /me:
    get:
      tags: [me]
      summary: Get Me
      operationId: get_me_me_get
      # Like POST /transactions: needs a bearer token, which an OpenAPI
      # example can't carry - so no inline examples;
      # specmatic:test covers it with externalized examples that send a real
      # token (see examples/get-me.json).
      security:
        - HTTPBearer: []
      responses:
        "200":
          description: The current user's profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              examples:
                unauthorized:
                  value:
                    detail: invalid or missing token
  /me/profile:
    put:
      tags: [me]
      summary: Update Profile
      operationId: update_profile_me_profile_put
      description: >-
        Changes the display name and/or language; fields left out stay as
        they are. The email is recorded but not editable here.
      security:
        - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: The updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              examples:
                unauthorized:
                  value:
                    detail: invalid or missing token
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /.well-known/jwks.json:
    get:
      tags: [auth]
      summary: Get Jwks
      description: The public key that verifies this API's access tokens (RS256 JWTs), as a JSON Web Key Set.
      operationId: get_jwks_well_known_jwks_json_get
      security: []
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jwks"
              examples:
                signing_key:
                  value:
                    keys:
                      - kty: RSA
                        use: sig
                        alg: RS256
                        kid: mXfbNPAfKC4MCud0
                        n: vG017uolFi2_VSJwcxfroKDhll4i_x31yScBe6nhL7l-L86Za6jGTkhrsNm4yv3O0RPGrA_eiFxM7OimL0c0vckCtP57fusQrnhcoAoqAFcRwtjHabee-G-mP3C5G46nj-1SudRpXFkEBYmjI3CNAad3MzQUFQ6kMtj8XkcK1KRJx0PRMdDLmZKpUhMIS5JnpMKnlNOQcL_dfQxHWjJrq8OKo2kYBrOl3JTKBWUTLPUWP22wAJUm_EODZ6GT1sUh82YRAsYSNaA9fKdm7qKeN7M7t2D8IEcG3QiG6iH0fvjMRrOjKeFzUWXwLk-AqtuJ6rRiVWDb_89A9Nl0Lhb1Jw
                        e: AQAB
  /transactions:
    get:
      tags: [transactions]
      summary: List Transactions
      operationId: list_transactions_transactions_get
      responses:
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                title: Response List Transactions Transactions Get
                items:
                  $ref: "#/components/schemas/TransactionOut"
    post:
      tags: [transactions]
      # Not "Create Account": quantity is a signed delta (see
      # TransactionCreate below), not an absolute value - this appends one
      # more event to an account's ledger (a journal entry), same wording
      # the frontend already uses for it (t.app.registerHeading, "Record a
      # Transaction" - see accounts/page.tsx). An account itself is never
      # created directly - GET /accounts is a read model derived from this
      # log, not a separate resource this operation writes to; a name with
      # no prior transactions is implicitly created by its first one (see
      # Transaction in models.py).
      summary: Record a Transaction
      operationId: create_transaction_transactions_post
      # Deliberately no requestBody/response `examples:` here, unlike every
      # other operation - this one needs a real bearer token (see auth.py),
      # which an OpenAPI example has no way to carry (it's a header, not
      # part of the body). specmatic:test already covers this operation via
      # its externalized example (examples/put-me-profile.json), with the
      # demo user's token;
      # an inline example here would give Specmatic a second, competing
      # scenario for the same operation with no way to authenticate it,
      # always failing with 401.
      security:
        - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionCreate"
      responses:
        "201":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionOut"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              examples:
                unauthorized:
                  value:
                    detail: invalid or missing token
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /transactions/{transaction_id}:
    get:
      tags: [transactions]
      summary: Get Transaction
      operationId: get_transaction_transactions__transaction_id__get
      parameters:
        - name: transaction_id
          in: path
          required: true
          schema:
            type: integer
            title: Transaction Id
      responses:
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionOut"
        "404":
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              examples:
                not_found:
                  value:
                    detail: transaction not found
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /accounts:
    get:
      tags: [accounts]
      # A read model derived from /transactions (one row per distinct name,
      # its balance the running sum of every transaction posted against
      # it) - not a separately created/stored entity, so this is the only
      # operation this resource has: no POST (see /transactions instead)
      # and no GET /accounts/{id}, since an account's identity is its name,
      # not an integer id the way a transaction's is.
      summary: List Accounts
      operationId: list_accounts_accounts_get
      responses:
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                title: Response List Accounts Accounts Get
                items:
                  $ref: "#/components/schemas/AccountBalance"
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AccountBalance:
      type: object
      title: AccountBalance
      description: >-
        A name's current balance: the sum of every event's `quantity` for
        that name, plus how many events contributed to it (see
        `account_service.get_balances`).
      required: [name, balance, eventCount]
      properties:
        name:
          type: string
          title: Name
        balance:
          type: integer
          title: Balance
        eventCount:
          type: integer
          title: Eventcount
    TransactionCreate:
      type: object
      title: TransactionCreate
      description: >-
        `quantity` is a signed delta (e.g. -3 to record consumption), not an
        absolute value - see `Transaction` in models.py.
      required: [name]
      properties:
        name:
          type: string
          title: Name
          minLength: 1
          maxLength: 128
        quantity:
          type: integer
          title: Quantity
          default: 0
          minimum: -2147483648
          maximum: 2147483647
    TransactionOut:
      type: object
      title: TransactionOut
      required: [id, name, quantity, source, createdAt]
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        quantity:
          type: integer
          title: Quantity
        source:
          type: string
          title: Source
        createdAt:
          type: string
          format: date-time
          title: Createdat
    HealthResponse:
      type: object
      title: HealthResponse
      required: [status]
      properties:
        status:
          type: string
          title: Status
    LoginRequest:
      type: object
      title: LoginRequest
      required: [username, password]
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
    Profile:
      type: object
      title: Profile
      description: >-
        The current user's profile (`GET /me`). `email` is recorded but not
        editable here: for a Keycloak user it is mirrored from the token, for
        a demo user it comes from the seed data.
      required: [username, language, provider]
      properties:
        username:
          type: string
          title: Username
        email:
          type: string
          nullable: true
          title: Email
        displayName:
          type: string
          nullable: true
          title: Displayname
        language:
          type: string
          enum: [ja, en]
          title: Language
        provider:
          type: string
          enum: [demo, keycloak]
          title: Provider
    ProfileUpdate:
      type: object
      title: ProfileUpdate
      description: "`PUT /me/profile` - only what the user may change."
      properties:
        displayName:
          type: string
          nullable: true
          maxLength: 64
          title: Displayname
        language:
          type: string
          enum: [ja, en]
          title: Language
    LoginResponse:
      type: object
      title: LoginResponse
      required: [token, username]
      properties:
        token:
          type: string
          title: Token
          description: "An RS256 JWT access token (send it as `Authorization: Bearer <token>`); verify it with /.well-known/jwks.json."
        username:
          type: string
          title: Username
    JwkKey:
      type: object
      title: JwkKey
      required: [kty, use, alg, kid, n, e]
      properties:
        kty: {type: string}
        use: {type: string}
        alg: {type: string}
        kid: {type: string}
        n: {type: string}
        e: {type: string}
    Jwks:
      type: object
      title: Jwks
      required: [keys]
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/JwkKey"
    HTTPError:
      type: object
      title: HTTPError
      description: FastAPI's default error body shape - HTTPException(detail=...).
      required: [detail]
      properties:
        detail:
          type: string
          title: Detail
    HTTPValidationError:
      type: object
      title: HTTPValidationError
      properties:
        detail:
          type: array
          title: Detail
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      title: ValidationError
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          title: Location
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context